src/main/java/com/thoughtworks/locker/SmartLockerRobot.java
package com.thoughtworks.locker;

import com.thoughtworks.locker.exception.FullCapacityException;

import java.util.List;

public class SmartLockerRobot {

    private final List<Locker> lockers;

    public SmartLockerRobot(List<Locker> lockers) {
        this.lockers = lockers;
    }

    public Ticket save(Bag bag) {
        return lockers.stream()
                .filter(Locker::isNotFull)
                .findFirst()
                .map(locker -> locker.save(bag))
                .orElseThrow(FullCapacityException::new);
    }
}
